--- Flutter.code-workspace
@@ -1,12 +1,0 @@
-{
-	"folders": [
-		{
-			"path": ".."
-		},
-		{
-			"name": "pirate_coin",
-			"path": "."
-		}
-	],
-	"settings": {}
-}
